plugins/requirements.txt
capstone
unicorn
keystone-engine
z3-solver
superglobals
setuptools
jsonschema
line_profiler
treelib
PyQt5
sortedcontainers